If VANADIUM_ROOT is defined, then use the go and js code there, otherwise fetch it ourselves.
For CI testing and debugging, we want the chat clients to build with the
code inside VANADIUM_ROOT.  That way we are always testing with
up-to-date code, and it makes debugging easier.

This CL causes Makefile to use the go and javascript code inside
VANADIUM_ROOT if it exists.  Otherwise it will download and install the
vanadium dependencies from the usual sources.

Also, I accidentally deleted a line from our .jshintignore during the
migration.  This CL puts it back.

Change-Id: Ic9561ed3fa96510bd3064a60239b64b4023ccdb5