Chat shell client: Get the members' names from their remote blessings.
We used to trust that chat clients would mount under their actual name.
That system would allow malicious clients to mount under a fake name,
and all other clients would be fooled.

This CL changes that.  After we glob on the channel name, we make a
dummy RPC to each server and get its blessings, which we know we
can trust.  The client's name is then derived from their blessings.

This CL also fixes a test that was accidentally not running.

Change-Id: If332f37fc431fee627900dbacfd8979e302a09f5