Chat web client: Get the members' names from their remote blessings.
We used to trust that chat clients would mount under their actual name.
That system would allow malicious clients to mount under a fake name,
and all other clients would be fooled.

This CL changes that.  After we glob on the channel name, we request the
remote blessings from each server, which we know we can trust.  The
client's name is then derived from their blessings.

There is a TODO to make sure that, when we send a message, the server we
bind to has the blessings that we expect.  This will prevent a malicious
server from mounting under the name of a member that just disconnected
and getting messages meant for the disconnected server.  We do this
check in Go, but the JS API does not support it yet.

This is the web-client equivalent of the shell-client CL in go/vcl/2942

Change-Id: I944df58da79fe225eaf9a475b764d6f5237f2745