Get rid of setTimeout(3000) before each web test.
Channel now emits a "ready" event once it has fetched the channel
members for the first time.  We listen for this event and only run the
tests once it has fired.

In order to make this work, I had to expose the "page" object on the
window, so the tests can access it.

Change-Id: If25f4c474abe5fcc2643e61f302f6292e14fb078