chat: Lock mount name so other users can't impersonate me.
Clients now pick a random mount name and attept to set an ACL on that
name that only allows them to Mount and Admin, but allows everybody else
to Resolve and Read.  Clients retry in the case that they pick a name
that is already locked.

When clients leave the channel, they Remove their name, which deletes
the ACL.

Change-Id: Ieda1664597c603f7a0b2e0e9cc1a5f0df2906f14