Chat cleanup part 1:
There is now a relatively sane way to run both the web and shell clients
in "dev mode" with a local mounttable and proxy.  That setup is
documented in the README.

You can now pass the mounttable and proxy to the web client, and the
shell client has a default proxy set.

I deleted a lot of dead tools and READMEs.  All documentation is now in
a single README, which will be easier to keep up-to-date.

Change-Id: Ib809b3b3a919e70c9ce5630fad182acabc7d38ff