chat: Change context mutators from SetFoo to WithFoo.
The motivation for this change is consistency, so our API becomes
easier to use.

Before this CL we were using the WithFoo style in the context
package, but the SetFoo style elsewhere, sometimes with an extra
"Context" thrown in for kicks.  After this CL we use WithFoo (and
GetFoo) everywhere.

Change-Id: I6c37f50f9c587f7df840eb8ac93e10ad8333f986
MultiPart: 4/4